break out student dashboard page into two components
